==> notebook/3. Feature Engineering.ipynb <==
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
%matplotlib inline

## Display all the columns of the dataframe
pd.pandas.set_option('display.max_columns',None)

import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)

earthquake_data = pd.read_csv('../datasets/final_dataset/cleaned_earthquake_data.csv')
earthquake_data.head()

earthquake_data['magnitude_type'].unique()

## Apply Feature Engineering

### Convert Categorical Feature into Numerical Feature

# 1. Ordinal encoding for magnitude_type because it is in order from least to major damage
from sklearn.preprocessing import OrdinalEncoder
from sklearn.preprocessing import StandardScaler, OrdinalEncoder

scaled_df = earthquake_data

ordinal_encoder = OrdinalEncoder()

# Ordinal Encoding
scaled_df['magnitude_type'] = ordinal_encoder.fit_transform(scaled_df[['magnitude_type']])

# Ordinal Encoding
scaled_df['day_name'] = ordinal_encoder.fit_transform(scaled_df[['day_name']])

scaled_df.head(5)

scaled_df['magnitude_type'].unique()

# Scaling continuous features
scaler = StandardScaler()

continious_feature_scale = ['latitude','longitude','depth', 'distance', 'gravity', 'force']
# Apply scaling
scaled_df[continious_feature_scale] = scaler.fit_transform(scaled_df[continious_feature_scale])

#### Cyclic encoding for time-related features

Time features often benefit more from encoding techniques that capture periodicity (like sine and cosine transformations) rather than standard scaling.

# Cyclic encoding for time-related features
scaled_df['month_sin'] = np.sin(2 * np.pi * scaled_df['month'] / 12)
scaled_df['month_cos'] = np.cos(2 * np.pi * scaled_df['month'] / 12)
scaled_df['hour_sin'] = np.sin(2 * np.pi * scaled_df['hour'] / 24)
scaled_df['hour_cos'] = np.cos(2 * np.pi * scaled_df['hour'] / 24)
scaled_df['minutes_sin'] = np.sin(2 * np.pi * scaled_df['minutes'] / 60)
scaled_df['minutes_cos'] = np.cos(2 * np.pi * scaled_df['minutes'] / 60)

scaled_df.head()

scaled_df['magnitude_type'].unique()

# Drop original month and hour columns after encoding
scaled_df = scaled_df.drop(columns=['month', 'hour','minutes'])

scaled_df.head(5)

# scaled_df.to_csv("../datasets/final_dataset/cyclic_scaled_data.csv", index=False)
