==> notebook/5. Split Data into sample.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
from datetime import datetime
import matplotlib.pyplot as plt
import plotly.express as px
import plotly.graph_objects as go

from sklearn.model_selection import train_test_split, cross_val_score, GridSearchCV, RandomizedSearchCV
from sklearn.ensemble import RandomForestRegressor, GradientBoostingRegressor, AdaBoostRegressor
from sklearn.linear_model import BayesianRidge, LinearRegression
from sklearn.preprocessing import PolynomialFeatures
from sklearn.pipeline import make_pipeline
from sklearn.tree import DecisionTreeRegressor
from xgboost import XGBRegressor
from catboost import CatBoostRegressor
from sklearn.neural_network import MLPRegressor
from sklearn.metrics import mean_squared_error, r2_score, mean_squared_error, mean_absolute_error
from sklearn.svm import SVR
from sklearn.linear_model import Ridge, Lasso, ElasticNet

import sys
from memory_profiler import profile

## Display all the columns of the dataframe
pd.pandas.set_option('display.max_columns',None)

import warnings

# Suppressing all warnings within the context
with warnings.catch_warnings():
    warnings.simplefilter("ignore")



image_path = f"../datasets/images"

final_df = pd.read_csv('../datasets/final_dataset/cyclic_scaled_data.csv')
final_df.head()

final_df.shape

#### Selected Feature After Feature Engineering

selected_independent_feature = ['latitude', 
                                'longitude',
                                'depth',
                                'magnitude',
                                'magnitude_type',
                                'distance',
                                'gravity',
                                'force',
                                'year',
                                'month_sin',
                                'month_cos']
print(len(selected_independent_feature))
final_df = final_df[selected_independent_feature]
final_df.head()

# final_df.to_csv("../datasets/final_dataset/total_data.csv", index=False)

### Select the Sample Data

    - Since your population size is 3,14,351, calculate the sizes for 25%, 50%, 80% and 100% samples:

# Determine the sample sizes
total_size = len(final_df)  # Total number of rows in the dataset
print("Total size: ",total_size)

sample_size_25 = int(0.25 * total_size)  #25% of the total data
sample_size_50 = int(0.50 * total_size)  # 50% of the total data
sample_size_65 = int(0.65 * total_size)  # 65% of the total data
sample_size_80 = int(0.80 * total_size)  # 80% of the total data

# Select the samples
sample_25 = final_df.sample(n=sample_size_25, random_state=42)
sample_50 = final_df.sample(n=sample_size_50, random_state=42)
sample_65 = final_df.sample(n=sample_size_65, random_state=42)
sample_80 = final_df.sample(n=sample_size_80, random_state=42)

print("sample_25: ", sample_25.shape)
print("sample_50: ", sample_50.shape)
print("sample_65: ", sample_65.shape)
print("sample_80: ", sample_80.shape)

# Save Sample data to file
# sample_25.to_csv("../datasets/final_dataset/sample_25.csv", index=False)
# sample_50.to_csv("../datasets/final_dataset/sample_50.csv", index=False)
# sample_65.to_csv("../datasets/final_dataset/sample_65.csv", index=False)
# sample_80.to_csv("../datasets/final_dataset/sample_80.csv", index=False)

df_sample = sample_25
# df_sample = sample_50
# df_sample = sample_65
# df_sample = sample_80

print("New Data Sample", df_sample.shape)
